SUKU midimonitor debugmonitor and luadebug cleanup
