+++ src/app/main/RightPanelContainer.svelte
@@ -4,7 +4,7 @@
 
   import Configuration from './panels/configuration/Configuration.svelte';
   import Preferences from './panels/preferences/Preferences.svelte';
-  import MidiMonitor from './panels/midi-monitor/MidiMonitor.svelte';
+  import MidiMonitor from './panels/MidiMonitor/MidiMonitor.svelte';
 
   export let classes;
