+++ src/app/main/panels/DebugMonitor/DebugMonitor.store.js
@@ -1,0 +1,34 @@
+import { writable, get } from 'svelte/store';
+
+
+function createDebugMonitor(){
+
+  const store = writable([]);
+
+  return {
+    ...store,
+    update_debugtext: (descr) => {
+
+      console.log(descr);
+
+      let sx = descr.brc_parameters.SX;
+      let sy = descr.brc_parameters.SY;
+      let text = descr.class_parameters.TEXT;
+
+
+      store.update(d => {
+       
+          if(d.length >= 15){
+            d.data.shift()
+          }
+          d = [...d, `[${sy},${sx}] ${text}`];
+        
+        return d;
+      })
+    }
+
+  }
+}
+
+
+export const debug_monitor_store = createDebugMonitor();
